+++ lib/counter_culture/counter.rb
@@ -1,6 +1,6 @@
 module CounterCulture
   class Counter
-    CONFIG_OPTIONS = [ :column_names, :counter_cache_name, :delta_column, :foreign_key_values, :touch, :delta_magnitude]
+    CONFIG_OPTIONS = [ :column_names, :counter_cache_name, :delta_column, :foreign_key_values, :touch, :delta_magnitude, :execute_after_commit ]
     ACTIVE_RECORD_VERSION = Gem.loaded_specs["activerecord"].version
 
     attr_reader :model, :relation, *CONFIG_OPTIONS
@@ -9,10 +9,6 @@ module CounterCulture
       @model = model
       @relation = relation.is_a?(Enumerable) ? relation : [relation]
 
-      if options.fetch(:execute_after_commit, false)
-        fail("execute_after_commit was removed; updates now run within the transaction")
-      end
-
       @counter_cache_name = options.fetch(:column_name, "#{model.name.demodulize.tableize}_count")
       @column_names = options[:column_names]
       @delta_column = options[:delta_column]
@@ -20,6 +16,18 @@ module CounterCulture
       @touch = options.fetch(:touch, false)
       @delta_magnitude = options[:delta_magnitude] || 1
       @with_papertrail = options.fetch(:with_papertrail, false)
+      @execute_after_commit = options.fetch(:execute_after_commit, false)
+
+      if @execute_after_commit
+        begin
+          require 'after_commit_action'
+        rescue LoadError
+          fail(LoadError.new(
+            "You need to include the `after_commit_action` gem in your "\
+            "gem dependencies to use the execute_after_commit option"))
+        end
+        model.include(AfterCommitAction)
+      end
     end
 
     # increments or decrements a counter cache
@@ -61,10 +69,16 @@ module CounterCulture
                           "#{model.connection.quote_column_name(change_counter_column)}"
                         end
 
+        column_type = klass.type_for_attribute(change_counter_column).type
+
         # we don't use Rails' update_counters because we support changing the timestamp
         updates = []
         # this updates the actual counter
-        updates << "#{quoted_column} = COALESCE(#{quoted_column}, 0) #{operator} #{delta_magnitude}"
+        if column_type == :money
+          updates << "#{quoted_column} = COALESCE(CAST(#{quoted_column} as NUMERIC), 0) #{operator} #{delta_magnitude}"
+        else
+          updates << "#{quoted_column} = COALESCE(#{quoted_column}, 0) #{operator} #{delta_magnitude}"
+        end
         # and here we update the timestamp, if so desired
         if touch
           current_time = obj.send(:current_time_from_proper_timezone)
@@ -93,14 +107,20 @@ module CounterCulture
                 instance.send("#{timestamp_column}=", current_time)
               end
 
-              instance.paper_trail.save_with_version(validate: false)
+              execute_now_or_after_commit(obj) do
+                instance.paper_trail.save_with_version(validate: false)
+              end
             else
-              instance.paper_trail.touch_with_version
+              execute_now_or_after_commit(obj) do
+                instance.paper_trail.touch_with_version
+              end
             end
           end
         end
 
-        klass.where(primary_key => id_to_change).update_all updates.join(', ')
+        execute_now_or_after_commit(obj) do
+          klass.where(primary_key => id_to_change).update_all updates.join(', ')
+        end
       end
     end
 
@@ -294,6 +314,16 @@ module CounterCulture
       end
 
       prev
+    end
+
+    def execute_now_or_after_commit(obj, &block)
+      execute_after_commit = @execute_after_commit.is_a?(Proc) ? @execute_after_commit.call : @execute_after_commit
+
+      if execute_after_commit
+        obj.execute_after_commit(&block)
+      else
+        block.call
+      end
     end
 
     private