+++ lib/counter_culture/reconciler.rb
@@ -74,7 +74,15 @@ module CounterCulture
 
         scope = relation_class
 
-        counter_column_names = column_names || {nil => counter_cache_name}
+        counter_column_names =
+          case column_names
+          when Proc
+            column_names.call
+          when Hash
+            column_names
+          else
+            { nil => counter_cache_name }
+          end
 
         if options[:column_name]
           counter_column_names = counter_column_names.select{ |_, v| options[:column_name].to_s == v }
@@ -187,7 +195,12 @@ module CounterCulture
         # if a delta column is provided use SUM, otherwise use COUNT
         return @count_select if @count_select
         if delta_column
-          @count_select = "SUM(COALESCE(#{self_table_name}.#{delta_column},0))"
+          # cast the column as NUMERIC if it is a PG money type
+          if model.type_for_attribute(delta_column).type == :money
+            @count_select = "SUM(COALESCE(CAST(#{self_table_name}.#{delta_column} as NUMERIC),0))"
+          else
+            @count_select = "SUM(COALESCE(#{self_table_name}.#{delta_column}, 0))"
+          end
         else
           @count_select = "COUNT(#{self_table_name}.#{model.primary_key})*#{delta_magnitude}"
         end