+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## 3.1.0 (November 29, 2021)
+
+Improvements:
+  - Allow specifiying a `Proc` to `column_names` to avoid loading a scope on
+    startup (#335)
+
+## 3.0.0 (October 26, 2021)
+
+Breaking changes:
+  - Dropped support for Ruby < 2.6
+  - Dropped support for Rails < 5.2
+
+Note that there are no specific breaking changes that would cause older
+versions of Ruby or Rails to stop working, we have simply stopped testing
+against them.
+
+Improvements:
+  - Support PostgreSql's `money` type for use with a delta column (#333)
+
+## 2.9.0 (August 27, 2021)
+
+Improvements:
+  - Allow `execute_after_commit` to be a `Proc` for dynamic control (#326)
+
+## 2.8.0 (March 16, 2021)
+
+Improvements:
+  - New `execute_after_commit` option (#309)
+
 ## 2.7.0 (November 16, 2020)
 
 Improvements: